Refactor : PostRequest
PostCreateRequest, PostUpdateRequest를 하나로 묶음
